Implemented NahrungDetails (first draft)

+++ app/src/main/java/com/example/stefansator/brealth/naehrstoffzentrale/Ingredient.java
@@ -19,7 +19,7 @@ import com.google.gson.annotations.SerializedName;
 public class Ingredient {
 
     @SerializedName("quantity")
-    private Integer quantity;
+    private Double quantity;
     @SerializedName("measureURI")
     private String measureURI;
     @SerializedName("foodURI")
@@ -27,19 +27,19 @@ public class Ingredient {
     @JsonIgnore
     private Map<String, Object> additionalProperties = new HashMap<String, Object>();
 
-    public Ingredient(Integer quantity, String measureURI, String foodURI) {
+    public Ingredient(Double quantity, String measureURI, String foodURI) {
         this.quantity = quantity;
         this.measureURI = measureURI;
         this.foodURI = foodURI;
     }
 
     @SerializedName("quantity")
-    public Integer getQuantity() {
+    public Double getQuantity() {
         return quantity;
     }
 
     @SerializedName("quantity")
-    public void setQuantity(Integer quantity) {
+    public void setQuantity(Double quantity) {
         this.quantity = quantity;
     }
